final Commit

+++ app/Http/Controllers/UserController.php
@@ -2,10 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\UserValidation;
 use App\Http\Resources\UserResource;
 use App\Models\Role;
 use App\Models\User;
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 use Inertia\Inertia;
 
 class UserController extends Controller
@@ -15,16 +16,16 @@ class UserController extends Controller
         return Inertia::render('User/Create');
     }
 
-    public function store(Request $request)
+    public function store(UserValidation $request)
     {
-        $data = $request->all();
-
-        $data['password'] = bcrypt('12345678');
-        $user = User::create($data);
-
-        $user->roles()->attach(
+        User::create([
+            'name' => $request->name,
+            'email' => $request->email,
+            'password' => bcrypt('12345678'),
+        ])->roles()->attach(
             Role::where('name', 'admin')->first()->id
-        );
+        ); 
+
         return redirect()->route('dashboard',[
             'type' => 'users'
         ]);
@@ -44,7 +45,7 @@ class UserController extends Controller
         ]);
     }
 
-    public function update(User $user, Request $request)
+    public function update(UserValidation $request, User $user)
     {
         $user->update($request->all());
 
@@ -55,10 +56,24 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        $validator = Validator::make(['id' => $user->id], ['id' => 'exists:users,id']);
+
+        $validator->after(function ($validator) use ($user) {
+            if ($user->employees()->exists()) {
+                $validator->errors()->add('error', "{$user->name} can't be deleted. First delete {$user->name}'s employees.");
+            }
+        });
+
+        if ($validator->fails()) {
+            return redirect()->route('dashboard', ['type' => 'users'])
+                            ->withErrors($validator);
+        }
+
         $user->roles()->detach();
 
         $user->delete();
 
-        back(303);
+        return redirect()->route('dashboard', ['type' => 'users']);
     }
+
 }
